Explicit install and compile extra deps
The xray propagator moved into the opentelemetry js contrib repo and
before it was outside, so we need to compile it there.

This new propagator aws xray lambda package is a test dependency of the
aws lambda instrumentation tests, so we aalso need to compile it there
otherwise the compilation of the lambda instrumentation fails.